Moved AuthenticationUtil.kt into authentication package.
