+++ Code/app/src/main/java/com/cmpt362/fitcheck/ui/authentication/AuthenticationUtil.kt
@@ -1,4 +1,4 @@
-package com.cmpt362.fitcheck.firebase
+package com.cmpt362.fitcheck.ui.authentication
 
 import android.util.Patterns
 import java.util.regex.Pattern
